+++ tests/subclass/secondary/c.py
@@ -7,7 +7,7 @@
 from ..base import BasePlugin
 
 
-class Trash(object): ...
+class Trash: ...
 
 
 TRASH_VAR = 1
@@ -20,4 +20,4 @@ class CPlugin(BasePlugin):
         return self.name
 
 
-class ATrash2(object): ...
+class ATrash2: ...