+++ tests/test_common.py
@@ -5,7 +5,6 @@
 # ---------------------------------------------------------------------
 
 # Python modules
-from typing import List, Tuple, Type
 
 # Third-party modules
 import pytest
@@ -16,49 +15,49 @@ from gufo.loader import Loader
 from .subclass.base import BasePlugin
 
 PLUGIN_BASES = ["tests.subclass.primary", "tests.subclass.secondary"]
-LoaderType = Loader[Type[BasePlugin]]
+LoaderType = Loader[type[BasePlugin]]
 
 
 @pytest.fixture(scope="module")
 def loader() -> LoaderType:
-    return Loader[Type[BasePlugin]](bases=PLUGIN_BASES)
+    return Loader[type[BasePlugin]](bases=PLUGIN_BASES)
 
 
 @pytest.fixture(scope="module")
 def exc_loader() -> LoaderType:
-    return Loader[Type[BasePlugin]](bases=PLUGIN_BASES, exclude=["d"])
+    return Loader[type[BasePlugin]](bases=PLUGIN_BASES, exclude=["d"])
 
 
 def test_base() -> None:
-    loader = Loader[Type[BasePlugin]](base=PLUGIN_BASES[0])
+    loader = Loader[type[BasePlugin]](base=PLUGIN_BASES[0])
     kls = loader["a"]
     assert issubclass(kls, BasePlugin)
 
 
 def test_no_base() -> None:
     with pytest.raises(RuntimeError):
-        Loader[Type[BasePlugin]]()
+        Loader[type[BasePlugin]]()
 
 
 def test_both_bases() -> None:
     with pytest.raises(RuntimeError):
-        Loader[Type[BasePlugin]](base=PLUGIN_BASES[0], bases=PLUGIN_BASES[1:])
+        Loader[type[BasePlugin]](base=PLUGIN_BASES[0], bases=PLUGIN_BASES[1:])
 
 
 def test_strict_miss() -> None:
     with pytest.raises(RuntimeError):
-        Loader[Type[BasePlugin]](
+        Loader[type[BasePlugin]](
             bases=[*PLUGIN_BASES, ".nonexistent"], strict=True
         )
 
 
 def test_no_strict_miss() -> None:
-    Loader[Type[BasePlugin]](bases=[*PLUGIN_BASES, ".nonexistent"])
+    Loader[type[BasePlugin]](bases=[*PLUGIN_BASES, ".nonexistent"])
 
 
 def test_no_paths() -> None:
     with pytest.raises(RuntimeError):
-        Loader[Type[BasePlugin]](bases=[".nonexistent1", ".nonexistent2"])
+        Loader[type[BasePlugin]](bases=[".nonexistent1", ".nonexistent2"])
 
 
 def test_get_none(loader: LoaderType) -> None:
@@ -84,7 +83,7 @@ def test_key_error(loader: LoaderType) -> None:
 
 
 def test_get_exclude() -> None:
-    loader = Loader[Type[BasePlugin]](bases=PLUGIN_BASES, exclude=["b"])
+    loader = Loader[type[BasePlugin]](bases=PLUGIN_BASES, exclude=["b"])
     kls = loader["a"]
     assert issubclass(kls, BasePlugin)
     with pytest.raises(RuntimeError):
@@ -92,7 +91,7 @@ def test_get_exclude() -> None:
 
 
 def test_cached_get() -> None:
-    loader = Loader[Type[BasePlugin]](bases=PLUGIN_BASES)
+    loader = Loader[type[BasePlugin]](bases=PLUGIN_BASES)
     kls1 = loader["a"]
     assert issubclass(kls1, BasePlugin)
     kls2 = loader["a"]
@@ -106,14 +105,14 @@ def test_keys(exc_loader: LoaderType) -> None:
 
 
 def test_values(exc_loader: LoaderType) -> None:
-    values: List[str] = []
+    values: list[str] = []
     for kls in exc_loader.values():
         values += [kls().get_name()]
     assert values == ["a", "b", "c"]
 
 
 def test_items(exc_loader: LoaderType) -> None:
-    items: List[Tuple[str, str]] = []
+    items: list[tuple[str, str]] = []
     for name, kls in exc_loader.items():
         items.append((name, kls().get_name()))
     assert items == [("a", "a"), ("b", "b"), ("c", "c")]